move into src/flagpole/*
